Added some plants

+++ MonoGame/Map/Plants/PlantConfig.cs
@@ -5,10 +5,10 @@ namespace MonoTest.Map.Plants
     public class PlantConfig
     {
         public Vector2 Position { get; }
-        public int AttackSpeed { get; }
+        public float AttackSpeed { get; }
         public int PlantType { get; }
 
-        public PlantConfig(Vector2 position, int attackSpeed, int plantType)
+        public PlantConfig(Vector2 position, float attackSpeed, int plantType)
         {
             Position = position;
             AttackSpeed = attackSpeed;
